split angles  to the separate file
