+++ Fitting/Mass/Kfunctions.py
@@ -1,0 +1,19 @@
+import sys
+from ROOT import RooRealVar, RooArgList
+from ROOT import RooVoigtian
+
+sys.path.append('../../') 
+from Naming.varDict import varDict
+
+var_dict = varDict()
+
+m_Kstar = RooRealVar(var_dict["Kstar_mass"],"m_{K*} [MeV]",846,946)
+
+v_mean = RooRealVar("v_mean", "Mean", 900, 880, 920)
+v_width = RooRealVar("v_width", "width", 50, 10, 100)
+v_sigma = RooRealVar("v_sigma", "Sigma", 15, 1, 120)
+
+voigt = RooVoigtian("voigt", "voigt", m_Kstar, v_mean, v_width, v_sigma)
+
+v_width.setVal(50)
+v_width.setConstant(True)
